+++ app/[locale]/projects/[projectId]/funding-search/page.tsx
@@ -38,7 +38,6 @@ const CreateProjectSearchFunder = () => {
   const trans = useTranslations('FunderSearch');
   const router = useRouter();
   const params = useParams();
-  const { projectId } = params;
 
   const [hasSearched, setHasSearched] = useState<boolean>(false);
   const [moreCounter, setMoreCounter] = useState<number>(0);
@@ -90,9 +89,7 @@ const CreateProjectSearchFunder = () => {
   }
 
   async function handleSelectFunder(funder: AffiliationSearch | FunderPopularityResult) {
-    const NEXT_URL = routePath('projects.create.projects.search', {
-      projectId: projectId as string,
-    });
+    const projectId = params.projectId as string;
     const input = {
       projectId: Number(projectId),
       affiliationId: funder.uri
@@ -104,7 +101,15 @@ const CreateProjectSearchFunder = () => {
         if (errs.length > 0) {
           setErrors(errs);
         } else {
-          router.push(NEXT_URL);
+          if (funder.apiTarget) {
+            router.push(routePath('projects.create.projects.search', {
+              projectId: projectId,
+            }));
+          } else {
+            router.push(routePath('projects.project.info', {
+              projectId: projectId,
+            }));
+          }
         }
       })
       .catch((err) => {
@@ -127,6 +132,7 @@ const CreateProjectSearchFunder = () => {
 
   function onResults(results: FunderSearchResults, isNew: boolean) {
     let validResults: AffiliationSearch[];
+
     if (results.items && results.items.length > 0) {
       validResults = results.items.filter((r): r is AffiliationSearch => r !== null)
     } else {
@@ -178,7 +184,7 @@ const CreateProjectSearchFunder = () => {
             moreTrigger={moreCounter}
           />
 
-          {popularFunders.length > 0 && (
+          {((popularFunders.length > 0) && !hasSearched)  && (
             <section aria-labelledby="popular-funders">
               <h3>{trans('popularTitle')}</h3>
               <div className={styles.popularFunders}>